fix: move shape generic implementations
